to_float needs to parse strings as doubles, not floats. (#2846)

+++ src/net/sourceforge/kolmafia/textui/RuntimeLibrary.java
@@ -4481,7 +4481,7 @@ public abstract class RuntimeLibrary {
     if (value.getType().equals(TypeSpec.STRING)) {
       String string = value.toString();
       try {
-        return new Value(StringUtilities.parseFloat(string));
+        return new Value(StringUtilities.parseDouble(string));
       } catch (NumberFormatException e) {
         Exception ex =
             controller.runtimeException(
